HV-508 fixing DecimalMaxValidatorForNumber and DecimalMinValidatorForNumber implementations

+++ hibernate-validator/src/main/java/org/hibernate/validator/constraints/impl/DecimalMaxValidatorForNumber.java
@@ -56,7 +56,7 @@ public class DecimalMaxValidatorForNumber implements ConstraintValidator<Decimal
 			return ( new BigDecimal( ( BigInteger ) value ) ).compareTo( maxValue ) != 1;
 		}
 		else {
-			return ( BigDecimal.valueOf( value.longValue() ).compareTo( maxValue ) ) != 1;
+			return ( BigDecimal.valueOf( value.doubleValue() ).compareTo( maxValue ) ) != 1;
 		}
 	}
 }
